Change memory and instruction to static

+++ CSCI_6461_GroupProject/src/Memory.java
@@ -4,19 +4,19 @@ public class Memory {
 	
 	//private variable need for the memory class
 	private Vector<Integer> addrVec;
-	private Vector<Integer> memoVec;
+	private Vector<Short> memoVec;
 	private Register registers;
 	
 	public Memory() {
 		addrVec = new Vector<Integer> (32);
-		memoVec = new Vector<Integer> (2048);
+		memoVec = new Vector<Short> (2048);
 		registers = new Register();
 		initRegister();
 	}
 	
 	public Vector<Integer> GetAddr() {return addrVec;} // Creating vector for address vectors
 
-	public Vector<Integer> GetMemo() {return memoVec;} // Creating vector for memory vectors
+	public Vector<Short> GetMemo() {return memoVec;} // Creating vector for memory vectors
 	
 	public Register GetRegister() { return registers; };
